Cleanup in exports, visibility and navigation
