+++ hacking-game/src/App.tsx
@@ -1,30 +1,27 @@
-import { useState } from "react";
-import { Routes, Route, Navigate } from "react-router-dom";
+import { useEffect, useState } from "react";
+import { Routes, Route, useNavigate } from "react-router-dom";
 import TerminalLayout from "./components/TerminalLayout";
 import IntroScreen from "./pages/IntroScreen";
 import HackingScreen from "./pages/HackingScreen";
 import FileExplorer from "./pages/FileExplorer";
 import FailureScreen from "./pages/FailureScreen";
-
-export enum GameResult {
-	Unresolved = 0,
-	Win = 1,
-	Lose = 2,
-}
+import { GameResult } from "./types/hacking";
 
 const App: React.FC = () => {
+	const navigate = useNavigate();
 	const [gameResult, setGameResult] = useState<GameResult>(GameResult.Unresolved);
 
-	const handleGameResult = (result: GameResult) => {
-		setGameResult(result);
-	};
+	useEffect(() => {
+		if (gameResult !== GameResult.Unresolved) {
+			navigate("/files");
+		}
+	}, [gameResult]);
 
 	return (
 		<Routes>
 			<Route path="/*" element={<TerminalLayout />}>
-				<Route path="game" element={<HackingScreen onResult={handleGameResult} />} />
-				<Route path="files" element={gameResult === GameResult.Win ? <FileExplorer /> : <Navigate to="/" />} />
-				<Route path="failure" element={<FailureScreen />} />
+				<Route path="game" element={<HackingScreen onResult={setGameResult} />} />
+				<Route path="files" element={gameResult === GameResult.Win ? <FileExplorer /> : <FailureScreen />} />
 				<Route path="" element={<IntroScreen />} />
 			</Route>
 		</Routes>
